api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
#azure-functions-worker
#azure-functions-durable
azure-cosmos
pytest
opencensus-ext-azure